About to do some potentially hazardous stuff
I'm added all the necessary seed data. I'm having trouble initializing the database. I think it has to do with how I implemented the Context classes. I think I should only have one, so I'm about to delete the other two. Hopefully, that's the issue, but if it's not, I can return to this commit.